+++ build.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
-# Recompile CSS files
-cd resources/DarkHumHub/less
-lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
-cd ../../DarkEnterprise/less
-lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
-cd ../../..
+# Recompile CSS files and extract colors
+cd resources/DarkHumHub
+lessc less/build.less css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
+css-color-extractor css/theme.css css/theme.css --format=css
+cd ../DarkEnterprise
+lessc less/build.less css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
+css-color-extractor css/theme.css css/theme.css --format=css
+cd ../../
 
 # Update message files
 cd ../../ php yii message/extract-module dark-mode